0-linear.c updated

+++ 0x1E-search_algorithms/0-linear.c
@@ -2,30 +2,25 @@
 
 /**
  * linear_search - searches for a value in an array
- * @array: pointer to the first element of the array
- * @size: number of elements in an array
+ * @array: first input of the array
+ * @size: size of the  array
  * @value: value to search for
  * Return: index of value || -1 if not found
  */
 
 int linear_search(int *array, size_t size, int value)
 {
-	int i = 0;
+	size_t i = 0;
 
-	/* Check if the input array is NULL */
 	if (array == NULL)
 		return (-1);
 
-	/* iterate through the array */
-	for (i = 0; i < (int)size; i++)
+	for (i = 0; i < size; i++)
 	{
-		/* print the comparison being made */
 		printf("Value checked array[%lu] = [%d]\n", i, array[i]);
 
-		/* if value is found return the index */
 		if (array[i] == value)
 			return (i);
 	}
-	/* Value not found in the array */
 	return (-1);
 }
